// src/store/main/system/types.ts
/*
 * @Author: korealu
 * @Date: 2022-02-09 09:56:39
 * @LastEditors: korealu
 * @LastEditTime: 2022-02-15 11:34:46
 * @Description: file content
 * @FilePath: /pofi-admin/src/store/main/system/types.ts
 */
export interface ISystemState {
  usersList: any[]
  usersCount: number
  rolesList: any[]
  rolesCount: number
  permissionsList: any[]
  permissionsCount: number
  routerList: any[]
  routerCount: number
  menuList: any[]
  menuCount: number
}